requirementx.txt
attrs==21.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
idna==3.2
pip==21.0.1
pycparser==2.20
pyOpenSSL==20.0.1
PySocks==1.7.1
requests==2.26.0
setuptools==52.0.0.post20210125
six==1.16.0
urllib3==1.26.6
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
